correct Error found by Travis

+++ java/src/jmri/jmrit/vsdecoder/EngineSoundEvent.java
@@ -168,7 +168,7 @@ public class EngineSoundEvent extends SoundEvent {
     @edu.umd.cs.findbugs.annotations.SuppressFBWarnings(value="BC_UNCONFIRMED_CAST", justification="DieselPane extends EnginePane")
     private void setDecoderVolume() {
         // Get decoder volume and pass it to the spinner
-        int dv = (Integer) Math.round(this.getParent().getDecoderVolume() * 100f);
+        int dv = Math.round(this.getParent().getDecoderVolume() * 100f);
         ((DieselPane) engine_pane).volume_slider.setValue(dv);
     }
